requirements.txt
requests==2.22.0
beautifulsoup4==4.8.0
pandas==0.25.1
python-dateutil==2.8.0
mysql-connector-python==8.0.18
